Final Result

+++ backend/src/main/java/com/putacovertoyourchild/app/controllers/AppUserController.java
@@ -3,6 +3,7 @@ package com.putacovertoyourchild.app.controllers;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,10 +11,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.putacovertoyourchild.app.exceptions.ValidationException;
 import com.putacovertoyourchild.app.entity.model.AppUser;
 import com.putacovertoyourchild.app.entity.service.IAppUserService;
 
 @RestController
+@CrossOrigin(origins = "http://localhost:8100")
 public class AppUserController {
 
 	@Autowired
@@ -25,10 +28,16 @@ public class AppUserController {
 	}
 	@GetMapping("/appuser/{id}")
 	public AppUser getOne(@PathVariable(value="id") long id) {
+		if (id <= 0) {
+			throw new ValidationException("El identificador no puede ser menor que 1");
+		}
 		return appUserService.getOne(id);
 	}
 	@PostMapping("/appuser")
 	public void post(AppUser appUser) {
+		if(appUser.getPrice().equals("0")) {
+			throw new ValidationException("El precio no puede ser 0");
+		}
 		appUserService.post(appUser);
 	}
 	@PutMapping("/appuser/{id}")
@@ -37,6 +46,9 @@ public class AppUserController {
 	}
 	@DeleteMapping("/appuser/{id}")
 	public void delete( @PathVariable(value="id") long id) {
+		if (id <= 0) {
+			throw new ValidationException("El identificador no puede ser menor que 1");
+		}
 		appUserService.delete(id);
 	}
